backend/requirements.txt
Flask==3.1.2
Flask-CORS==6.0.0
fiftyone==1.9.0
Pillow==12.0.0
python-dotenv==1.0.0
gunicorn==21.2.0
gdown==5.2.1
azure-cosmos
azure-identity
azure-storage-blob
opencv-python
# Additional recommended dependencies
pymongo>=4.6.0
requests>=2.31.0
Flask-RESTful>=0.3.9
Werkzeug>=2.3.0
numpy>=1.24.0
scikit-learn>=1.3.0
pytest>=7.4.0
